[fix] redisson: Duration to int 변환 시 오버플로우 방지 로직 추가
- Duration.toMillis()를 int로 강제 형변환 시 오버플로우 위험(잘못된 설정값(예: 30일)이 입력되면 음수로 변환되어 오작동 가능)
- 현재 설정값은 모두 int 범위 내이지만, 향후 잘못된 설정 입력을 방지하기 위한 방어 코드 추가
- 해결:
  - (int) 강제 형변환 -> Math.toIntExact() 사용
  - 오버플로우 발생 시 ArithmeticException 발생
  - 애플리케이션 시작 시점에 문제 조기 발견 가능

+++ src/main/java/org/example/ootoutfitoftoday/common/config/RedissonConfig.java
@@ -45,15 +45,16 @@ public class RedissonConfig {
                 // 커넥션 풀 설정
                 .setConnectionPoolSize(redisProperties.getLettuce().getPool().getMaxActive())
                 .setConnectionMinimumIdleSize(redisProperties.getLettuce().getPool().getMinIdle())
-                // Redisson 전용 설정
-                .setIdleConnectionTimeout((int) redissonProperties.getIdleConnectionTimeout().toMillis())
+                // Redisson 전용 설정(Duration -> int 안전 변환)
+                // Math.toIntExact(): 오버플로우 발생 시 ArithmeticException 발생
+                .setIdleConnectionTimeout(Math.toIntExact(redissonProperties.getIdleConnectionTimeout().toMillis()))
                 // 타임아웃 설정
-                .setConnectTimeout((int) redisProperties.getTimeout().toMillis())
-                .setTimeout((int) redisProperties.getTimeout().toMillis())
-                // 재시도 설정
+                .setConnectTimeout(Math.toIntExact(redisProperties.getTimeout().toMillis()))
+                .setTimeout(Math.toIntExact(redisProperties.getTimeout().toMillis()))
+                // 재시도 설정(Duration -> int 안전 변환)
                 .setRetryAttempts(redissonProperties.getRetryAttempts())
                 // deprecated 되었으나, 신규의 setRetryInterval(Duration) 사용 불가
-                .setRetryInterval((int) redissonProperties.getRetryInterval().toMillis());
+                .setRetryInterval(Math.toIntExact(redissonProperties.getRetryInterval().toMillis()));
 
         // 비밀번호 설정
         if (StringUtils.hasText(redisProperties.getPassword())) {
